requirements.txt
streamlit
pandas
Pillow
qrcode
